aws: instance object cache for instance dupe prevention
Multiple physical instances may be created when they all should
end up on a single VM instance when the AWS server instance listing
is not immediately updated to reflect the newly created instances.
This seems to happen rarely but still sometimes.

The workaround is to maintain client-side bookkeeping of the
created instances so that the same instance is always re-used
when necessary.