cloud_vsphere: the 'pristine' snapshot can be anywhere, not only the fist one
The VM status check and revert logic required the base snapshot to be called
'pristine' and be the first one on the list. With this change the snapshot
can be located anywhere in the snapshot tree as long as it has the right name.
Consequentially there can be other, even never snapshots too, they will not
disturb the revert, as long as there is only one named 'pristine'.