cloud_libvirt: support for recent paramiko versions
Paramiko v1.8.0-11-g31ea4f0 added support for tunneled connections using a
new `sock` argument to `connect()` instead of a new `connect_socket()`
function which Poni expected (which was based on my pull request at
https://github.com/paramiko/paramiko/pull/39)

Poni monkey-patched Paramiko to add `connect_socket()` if it was missing,
but that got broken by Paramiko v1.12.1-8-g3e1f9f0 which added new GSS
related arguments to the internal `_auth()` function.

Fix this by using the new `sock` parameter when available and by
monkey-patching that in when it's missing.